requirements.txt
nfcpy
ndeflib
cryptography
paho-mqtt
pyyaml
argparse
